2059: Squash classes
