+++ app/Stats.php
@@ -17,8 +17,11 @@ declare(strict_types=1);
 
 namespace Fisharebest\Webtrees;
 
-use Fisharebest\Webtrees\Statistics\Birth;
-use Fisharebest\Webtrees\Statistics\Death;
+use Fisharebest\Webtrees\Functions\FunctionsDate;
+use Fisharebest\Webtrees\Module\ModuleBlockInterface;
+use Fisharebest\Webtrees\Module\ModuleInterface;
+use Fisharebest\Webtrees\Statistics\BirthPlaces;
+use Fisharebest\Webtrees\Statistics\DeathPlaces;
 use Fisharebest\Webtrees\Statistics\Google;
 use Fisharebest\Webtrees\Statistics\Helper\Country;
 use Fisharebest\Webtrees\Statistics\Helper\Sql;
@@ -1293,7 +1296,7 @@ class Stats implements
      */
     public function statsBirthQuery($simple = true, $sex = false, $year1 = -1, $year2 = -1): array
     {
-        return (new Birth($this->tree))->query($sex, $year1, $year2);
+        return $this->individualRepository->statsBirthQuery($sex, $year1, $year2);
     }
 
     /**
@@ -1307,8 +1310,7 @@ class Stats implements
      */
     public function statsBirth(string $size = null, string $color_from = null, string $color_to = null): string
     {
-        return (new Google\ChartBirth($this->tree))
-            ->chartBirth($size, $color_from, $color_to);
+        return $this->individualRepository->statsBirth($size, $color_from, $color_to);
     }
 
     /**
@@ -1403,7 +1405,7 @@ class Stats implements
      */
     public function statsDeathQuery($simple = true, $sex = false, $year1 = -1, $year2 = -1): array
     {
-        return (new Death($this->tree))->query($sex, $year1, $year2);
+        return $this->individualRepository->statsDeathQuery($sex, $year1, $year2);
     }
 
     /**
@@ -1417,8 +1419,7 @@ class Stats implements
      */
     public function statsDeath(string $size = null, string $color_from = null, string $color_to = null): string
     {
-        return (new Google\ChartDeath($this->tree))
-            ->chartDeath($size, $color_from, $color_to);
+        return $this->individualRepository->statsDeath($size, $color_from, $color_to);
     }
 
     /**
@@ -1720,7 +1721,7 @@ class Stats implements
      */
     public function statsAgeQuery($simple = true, $related = 'BIRT', $sex = 'BOTH', $year1 = -1, $year2 = -1)
     {
-        return (new StatisticAge($this->tree))->query($related, $sex, $year1, $year2);
+        return $this->individualRepository->statsAgeQuery($related, $sex, $year1, $year2);
     }
 
     /**
@@ -1732,7 +1733,7 @@ class Stats implements
      */
     public function statsAge(string $size = '230x250'): string
     {
-        return (new Google\ChartAge($this->tree))->chartAge($size);
+        return $this->individualRepository->statsAge($size);
     }
 
     /**
